refactor: follow dotnet format suggestions
