+++ Content/src/Modules/MainModule.cs
@@ -1,6 +1,6 @@
 ﻿using System.Threading.Tasks;
 using Carter;
-using CarterService.Entities;
+using CarterService.Entities.Internal;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Routing;
