+++ Content/src/Repositories/IHelloRepository.cs
@@ -1,4 +1,4 @@
-﻿namespace CarterService.Repository;
+﻿namespace CarterService.Repositories;
 
 public interface IHelloRepository
 {
